refactor(web): extract useBranchModifier hook from ChatInput
ChatInput.tsx is a 650+ line component. The branch-modifier tracking
(a useState + a window keydown/keyup/blur listener effect) is fully
self-contained — it reads no props and touches no other local state —
so move it into lib/hooks/useBranchModifier.ts. Pure code motion, no
behavior change.
